Json-Generator: Show all strings with pretty print
The Json-Generator filtered untranslated string out for the JSON-File.
With such file a developer needs more time to translate things
because he needs to look it up several times.

For a better workflow all translatable strings are included in the JSON-File.
With pretty-print (>= PHP 5.4) the JSON is more human readable.