Fixed JavaScript source extraction with single quoted functions (#141)
* JavaScript function scanner simple quote fix

Fix for scanning JS source codes with special tags in simple quoted messages. For example:
i18n.__(‘This is a <a action=“javascript:location.reload()">test</a>')

* JavaScript function scanner simple quote fix

Fix for scanning JS source codes with special tags in simple quoted messages. For example:
i18n.__(‘This is a <a href=“javascript:location.reload()">test</a>')